+++ core/CHANGELOG.md
@@ -21,6 +21,8 @@
 * Allow configurable, custom Docker environment variables that will be passed as-is to the test suite
 * Added `--list` arg to print test names in test suite
 * Kill unnecessary `TestSuiteRunner`, `TestExecutorParallelizer`, and `TestExecutor` structs
+* Modify the machinery to only use a single Docker volume for an entire test suite execution
+* Containerize the Docker initializer
 
 # 0.9.0
 * Change ConfigurationID to be a string